# Sarah.Server/Models/Dtos/RoomDto.cs
namespace Sarah.Server.Models.Dtos
{
    public class RoomDto
    {
        public long Id { get; set; }
        public string Name { get; set; } = "";
        public bool IsFavourite { get; set; }
    }
}