# sarah.client/src/app/services/api-client/model/roomDto.ts
/**
 * Sarah.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RoomDto { 
    id?: number;
    name?: string | null;
    isFavourite?: boolean;
}